thursday 28.5.2026 second part

+++ part4/bloglist-backend/controllers/blogs.js
@@ -2,10 +2,13 @@
 
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const middleware = require('../utils/middleware')
+const userExtractor = middleware.userExtractor
 // Exercise 4.17 Blog List Expansions, step 5
-const User = require('../models/user')
+//const User = require('../models/user')
 // Exercise 4.19: Blog List Expansions, step 7
-const jwt = require('jsonwebtoken')
+//const jwt = require('jsonwebtoken')
+
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', { username: 1, name: 1, id: 1 }) // Exercise 4.17 Blog List Expansions, step 5
@@ -13,26 +16,27 @@ blogsRouter.get('/', async (request, response) => {
 })
 
 // Exercise 4.19 Blog List Expansions, step 7
-const getTokenFrom = request => {
+/*const getTokenFrom = request => {
   const authorization = request.get('authorization')
   if (authorization && authorization.startsWith('Bearer ')) {
     return authorization.replace('Bearer ', '')
   }
   return null
 }
+  */
 
 
-blogsRouter.post('/', async (request, response) => {
+blogsRouter.post('/', userExtractor, async (request, response) => {
   const body = request.body
 
   // Exercise 4.19 Blog List Expansions, step 7
-  const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
+  /*const decodedToken = jwt.verify(request.token, process.env.SECRET)
   if (!decodedToken.id) {
     return response.status(401).json({ error: 'token invalid' })
-  }
+  }*/
 
   // Exercise 4.17 Blog List Expansions, step 5
-  const user = await User.findById(decodedToken.id)
+  const user = request.user // Exercise 4.22 Blog List Expansion, step 10
   if (!user) {
     return response.status(400).json({ error: 'userId missing or not valid' })
   }
@@ -53,8 +57,18 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 // Exercise 4.13: Blog List Expansions, step 1
-blogsRouter.delete('/:id', async (request, response) => {
+blogsRouter.delete('/:id', userExtractor, async (request, response) => {
+
+  // Exercise 4.21 Blog List Expansion, step 9
+  const blogToDelete = await Blog.findById(request.params.id)
+  // Exercise 4.22 Blog List Expansion, step 10
+  const user = request.user
+  if (blogToDelete.user.toString() !== user.id.toString()) {
+    return response.status(403).json({ error: 'only the creator of the blog can delete it' })
+  }
   await Blog.findByIdAndDelete(request.params.id)
+  user.blogs = user.blogs.filter(b => b.id.toString() !== blogToDelete.id.toString())
+  await user.save()
   response.status(204).end()
 })
